Return an unhandled exceptions in callback
If there was an unhandled exception, invoking the callback from the extension
 will return null.  Add a field to the map wrapper called "threw_exception" that
gets initialized to false before every call to s.circle() or s.beam().  If the
global pyfov handler ever detects that NULL was returned by the assigned python
callback, then set this.  When execution return to the original call to pyfov_beam
or pyfov_circle, return NULL, so the most recent exception will get thrown by
Python